\name{plot.somgrid}
\alias{plot.somgrid}
\title{Plot a Self-Organising Map prior structure}
\description{
Plot the prior structure used by a Self-Organizing Map.  The rendering
is glyph based.  Rectangular grids use square glyphs while hexagonal
grids use hexagons (the generic term is cell).  Glyph positioning is
fixed (according to the prior structure) but glyph size and color can be
used to encode different type of results. 
}
\usage{
\method{plot}{somgrid}(x, col = NA, size = NA, add = FALSE, border =
   NULL,colorValues = NA, sizeValues = NA,
   colormap = 50,withkey=FALSE, ...)
}
\arguments{
  \item{x}{an object of class \code{somgrid}}
  \item{col}{colors of the cells}
  \item{size}{sizes of the cells (must be between 0 and 1, inclusive)}
  \item{add}{logical. If \code{TRUE}, add to a current plot}
  \item{border}{the color to draw the border. The default, \code{NULL}, means to use \code{\link{par}}("fg"). Use \code{border = NA} to omit borders.}
  \item{colorValues}{a vector containing the values to be used for calculating the colors of the cells. If specified, this takes precedence over the \code{col} parameter. The color mapping is linear.}
  \item{sizeValues}{a vector containing the values to be used for calculating the size of the cells. If specified, this takes precedence over the \code{size} parameter. The surface of each cell will be proportional to the square root of the corresponding value.}
  \item{colormap}{if an integer, gives the number of colors to be generated by \code{\link{heat.colors}} for the coloring of the cells with \code{colorValues}. Alternatively, a vector of colors to use in conjunction with \code{colorValues}}
  \item{withkey}{logical. If \code{TRUE} add a key to the plot (incompatible with \code{add=TRUE}). See details}
  \item{\dots}{additional parameters transmitted to the low level plot function (i.e., \code{\link{polygon}} for hexagonal grids and \code{\link{rect}} for rectangular grids)}
}
\details{
When asked for a key, this function uses the \code{\link{layout}} facilities. It is therefore incompatible with other device splitting functions (\code{\link{par}}(mfrow), \code{par(mfcol)} and \code{\link{split.screen}}).
}
\references{
  Vesanto, J. (2002) \emph{Data Exploration Process Based on the
    Self--Organizing Map}, PhD thesis, Helsinki University of Technology, Espoo (Finland), Acta Polytechnica Scandinavica, Mathematics and Computing Series No.
  115.
}
\author{Fabrice Rossi}
%\note{
%}
%\seealso{ ~~objects to See Also as \code{\link{help}}, ~~~ }
\examples{
data(iris)
# scaling
data <- scale(iris[1:4])

# a medium hexagonal grid
sg <- somgrid(xdim=15,ydim=15,topo="hex")

# empty grid
plot(sg)

# choose a good SOM
som <- som.tune(data,sg)$best.som

# color based on the first coordinate (sepal length), size on the second
# (sepal width) of the prototypes
plot(sg,colorValues=som$prototypes[,1],sizeValues=som$prototypes[,2],
     withkey=TRUE)

# color based on the dominant class
col <- rep(NA,sg$size)
classes <- sapply(mapToUnit(som,iris[[5]]),summary)
col[colSums(classes)!=0] <-
    c("red","blue","green")[apply(classes[,colSums(classes)!=0],2,which.max)]
plot(sg,col=col)
}
\keyword{cluster}
\keyword{hplot}
