\name{print.somgrid}
\alias{print.somgrid}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Print method for 'somgrid' objects}
\description{
Describe the prior structure given by a \code{somgrid} object.
}
\usage{
\method{print}{somgrid}(x,...)
}
%- maybe also \code{usage} for other objects documented here.
\arguments{
  \item{x}{a \code{somgrid} object}
  \item{...}{not used}
}
\author{Fabrice Rossi}
% Add one or more standard keywords, see file \CODE{KEYWORDS} in the
% R documentation directory.
\keyword{cluster}
\keyword{print}
