\name{summary.somgrid}
\alias{summary.somgrid}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Summary method for 'somgrid' objects}
\description{
Describe the prior structure given by a \code{somgrid} object.
}
\usage{
\method{summary}{somgrid}(object,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{a \code{somgrid} object}
  \item{...}{not used}
}
\author{Fabrice Rossi}
%\seealso{}
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{cluster}
\keyword{print}
